Replace paragraph-count <a> tags by pseudo-elements
I think we'll be happier in the long run not having
to generate/write inline tags for every paragraph
and use CSS counters instead.

The trade-off is we lose the ability to click on
the paragraph number to get that precise URL, but 
the URL is still updated on scroll, so that's not
too bad.

The current behavior was conflicting with
the scroll actually, so that's one bug less to fix.
:see_no_evil: